+++ Readers/IReader.cs
@@ -20,8 +20,8 @@ namespace PointCloudConverter.Readers
 
         // close filestream
         void Close();
-        float GetIntensity();
-        float GetClassification();
+        byte GetIntensity();
+        byte GetClassification();
         LasHeader GetMetaData(ImportSettings importSettings, int fileIndex);
         void Dispose();
     }
